+++ aioxmpp/pep/service.py
@@ -32,6 +32,17 @@ class PEPClient(service.Service):
     """
     :class:`PEPClient` simplifies working with PEP services.
 
+    .. versionchanged:: 0.10
+
+        Before version 0.10, this service did not depend on
+        :class:`aioxmpp.EntityCapsService`. This was surprising to users because
+        the PEP relies on the functionality provided by the protocols
+        implemented by :class:`~aioxmpp.EntityCapsService` to provide key
+        features. With the release 0.10, :class:`~aioxmpp.EntityCapsService` is
+        a dependency of this service. For backward compatibility, your
+        application should still :meth:`aioxmpp.Client.summon` the
+        :class:`~aioxmpp.EntityCapsService` explicitly.
+
     Compared to :class:`~aioxmpp.PubSubClient` it supports automatic
     checking for server support, a stream-lined API. It is intended to
     make PEP things easy. If you need more fine-grained control or do
@@ -59,6 +70,7 @@ class PEPClient(service.Service):
         aioxmpp.DiscoClient,
         aioxmpp.DiscoServer,
         aioxmpp.PubSubClient,
+        aioxmpp.EntityCapsService,
     ]
 
     def __init__(self, client, **kwargs):